requirements.txt
aiogram==3.22.0
aiomysql==0.2.0
sqlalchemy==2.0.27
alembic==1.13.1
APScheduler==3.10.4
pytz==2024.1
httpx==0.26.0
pydantic==2.6.1
pydantic-settings==2.1.0
python-docx==1.1.0
reportlab==4.0.9
loguru==0.7.2
python-dotenv==1.0.0
